@@ -14,19 +14,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.0",
+    "@angular/animations": "^5.0.0",
     "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.3.0",
-    "@angular/compiler": "^4.3.0",
-    "@angular/core": "^4.3.0",
-    "@angular/forms": "^4.3.0",
-    "@angular/http": "^4.3.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
     "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.3.0",
-    "@angular/platform-browser-dynamic": "^4.3.0",
-    "@angular/router": "^4.3.0",
-    "@ngx-translate/core": "^7.0.0",
-    "@ngx-translate/http-loader": "^0.1.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
+    "@ngx-progressbar/core": "^4.1.0",
+    "@ngx-translate/core": "^9.1.1",
+    "@ngx-translate/http-loader": "^2.0.1",
     "core-js": "^2.4.1",
     "flot": "^0.8.0-alpha",
     "jquery-table-fixed-header": "^0.1.1",
@@ -34,14 +36,14 @@
     "materialize-css": "^0.98.2",
     "mdi": "^1.9.33",
     "ng2-materialize": "1.5.1",
-    "ngx-progressbar": "^2.1.1",
-    "rxjs": "^5.1.0",
+    "rxjs": "^5.5.2",
     "tinymce": "^4.6.4",
+    "webpack-sources": "^1.1.0",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.2",
-    "@angular/compiler-cli": "^4.3.0",
+    "@angular/cli": "^1.6.8",
+    "@angular/compiler-cli": "^5.0.0",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
     "codelyzer": "~2.0.0",
@@ -50,12 +52,12 @@
     "karma": "~1.4.1",
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "typescript": "2.4.2"
   }
 }
