@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
-import { Router, ActivatedRoute, Params, NavigationStart, NavigationEnd, NavigationExtras } from '@angular/router';
+import { Router, ActivatedRoute, NavigationExtras } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 import { Observable } from 'rxjs/Rx';
 import { Subject } from 'rxjs/Subject';
 
-import { Http, Response, URLSearchParams, Headers, RequestOptions } from '@angular/http';
+import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
 
 import { AppState } from './app.state';
 import { ListValue } from './models/list-value';
@@ -26,7 +26,7 @@ export class AppService {
     private translate: TranslateService,
     private router: Router,
     private route: ActivatedRoute,
-    private http: Http) { }
+    private http: HttpClient) { }
 
 
   changeLang(lang: string) {
@@ -62,68 +62,68 @@ export class AppService {
     }
   }
   
-  doHomeSearchParams(boosted: boolean = true): URLSearchParams {
+  doHomeSearchParams(boosted: boolean = true): HttpParams  {
     
-    let params: URLSearchParams = new URLSearchParams();
-    params.set('q', '*');
-    params.set('facet', 'true');
-    params.set('facet.mincount', '1');
+    let params: HttpParams = new HttpParams();
+    params = params.set('q', '*');
+    params = params.set('facet', 'true');
+    params = params.set('facet.mincount', '1');
     for (let i in this.state.config['facets']) {
-        params.append('facet.field', this.state.config['facets'][i]['field']);
+        params = params.append('facet.field', this.state.config['facets'][i]['field']);
     }
-    params.set('facet.range', 'rokvyd');
-    params.set('facet.range.start', '1');
-    params.set('facet.range.end', (new Date()).getFullYear() + '');
-    params.set('facet.range.gap', '10');
+    params = params.set('facet.range', 'rokvyd');
+    params = params.set('facet.range.start', '1');
+    params = params.set('facet.range.end', (new Date()).getFullYear() + '');
+    params = params.set('facet.range.gap', '10');
 
-    params.set('rows', '0');
+    params = params.set('rows', '0');
     this.state.clearParams();
     return params
   }
 
-  doSearchParams(sroute: string, boosted: boolean = true, mm: string = '100%'): URLSearchParams {
+  doSearchParams(sroute: string, boosted: boolean = true, mm: string = '100%'): HttpParams {
     
     if(sroute === 'home'){
       return this.doHomeSearchParams(false);
     }
-    let params: URLSearchParams = new URLSearchParams();
+    let params: HttpParams = new HttpParams();
     if (this.state.q && this.state.q !== '') {
-      params.set('q', this.state.q);
+      params = params.set('q', this.state.q);
     } else {
-      params.set('q', '*');
+      params = params.set('q', '*');
     }
 
-    params.set('start', this.state.start + '');
-    params.set('rows', this.state.rows + '');
-    params.set('sort', this.state.currentSort.field);
-    params.set('facet', 'true');
-    params.set('facet.mincount', '1');
+    params = params.set('start', this.state.start + '');
+    params = params.set('rows', this.state.rows + '');
+    params = params.set('sort', this.state.currentSort.field);
+    params = params.set('facet', 'true');
+    params = params.set('facet.mincount', '1');
     for (let i in this.state.config['facets']) {
-      params.append('facet.field', this.state.config['facets'][i]['field']);
+      params = params.append('facet.field', this.state.config['facets'][i]['field']);
     }
 
-    params.set('facet.range', '{!ex=rk}rokvyd');
+    params = params.set('facet.range', '{!ex=rk}rokvyd');
     
 //    params.set('facet.range.start', this.state.currentOd + '');
 //    params.set('facet.range.end', this.state.currentDo + '');
     
-    params.set('facet.range.start', '0');
-    params.set('facet.range.end', (new Date()).getFullYear() + '');
-    params.set('facet.range.gap', '10');
+    params = params.set('facet.range.start', '0');
+    params = params.set('facet.range.end', (new Date()).getFullYear() + '');
+    params = params.set('facet.range.gap', '10');
     
     if(boosted){
-      params.set('qf', this.state.config['boost']['qf']);
+      params = params.set('qf', this.state.config['boost']['qf']);
     }
-    params.set('mm', mm);
+    params = params.set('mm', mm);
 
       
     for (let i in this.state.usedFilters) {
       if(this.state.usedFilters[i].field === 'rokvyd'){
         //let vals = this.state.usedFilters[i].value
         let fq = '[' + this.state.currentOd + ' TO ' + this.state.currentDo + '}';
-        params.append('fq', '{!tag=rk}rokvyd:' + fq);
+        params = params.append('fq', '{!tag=rk}rokvyd:' + fq);
       } else {
-        params.append('fq', this.state.usedFilters[i].field + ':"' + this.state.usedFilters[i].value + '"');
+        params = params.append('fq', this.state.usedFilters[i].field + ':"' + this.state.usedFilters[i].value + '"');
       }
     }
     
@@ -132,7 +132,7 @@ export class AppService {
       if(this.state.advParams[i] !== null && this.state.advParams[i] !== ''){
         if(i === 'title'){
           
-          params.append('fq', 'title:"' + this.state.advParams[i].trim() +
+          params = params.append('fq', 'title:"' + this.state.advParams[i].trim() +
            '" OR title_prefix:"' + this.state.advParams[i].trim() +
            '" OR varnazev:"' + this.state.advParams[i].trim() + '"');
           if (boosted && (!this.state.q || this.state.q === '')) {
@@ -141,19 +141,19 @@ export class AppService {
 //              '"^2.0 varnazev:"' + this.state.advParams[i].trim() +
 //               '"^1.1 title_prefix:"' + this.state.advParams[i].trim() +
 //                '"^4.0 title:"' + this.state.advParams[i].trim() + '"^1.5 ');
-            params.set('q', this.state.advParams[i].trim());
+            params = params.set('q', this.state.advParams[i].trim());
                 
           }
         }else{
-          params.append('fq', i + ':"' + this.state.advParams[i].trim() + '"');
+          params = params.append('fq', i + ':"' + this.state.advParams[i].trim() + '"');
         }
       }
     }
 
     if (this.state.currentCollapse['field'] !== 'id') {
-      params.append('fq', '{!collapse field=' + this.state.currentCollapse['field'] + '}');
-      params.append('expand', 'true');
-      params.append('expand.rows', '1');
+      params = params.append('fq', '{!collapse field=' + this.state.currentCollapse['field'] + '}');
+      params = params.append('expand', 'true');
+      params = params.append('expand.rows', '1');
     }
     
     return params;
@@ -183,40 +183,32 @@ export class AppService {
     return params;
   }
 
-  search(params: URLSearchParams, type: string = 'results'): void {
+  search(params: HttpParams, type: string = 'results'): Observable<Object> {
     this.state.startSearch(type);
     var url = this.state.config['context'] + 'search/rdcz/select';
-    this.http.get(url, { search: params }).map(res => {
-      this.state.processSearch(res.json(), type);
-    }).subscribe();
+    return this.http.get(url, { params: params });
   }
   
-  getDigObjects(params: URLSearchParams){
+  getDigObjects(params: HttpParams){
     var url = this.state.config['context'] + 'search/digobjekt/select';
-    return this.http.get(url, { search: params }).map((res: Response) => {
-      return res.json();
-    });
+    return this.http.get(url, { params: params });
     
   }
   
   searchAleph(){
     var url = this.state.config['context'] + 'aleph';
-    let params: URLSearchParams = new URLSearchParams();
+    let params: HttpParams = new HttpParams();
     params.set('bc', this.state.q)
-    return this.http.get(url, { search: params }).map((res: Response) => {
-      
-      return res.json();
-      //return res.text();
-    });
+    return this.http.get(url, { params: params });
   }
   
   searchAlephDirect(){
     var url = this.state.config['context'] + 'alephDirect';
-    let params: URLSearchParams = new URLSearchParams();
+    let params: HttpParams = new HttpParams();
     params.set('base', 'nkc');
     params.set('op', 'find');
     params.set('request', 'bc='+this.state.q);
-    return this.http.get(url, { search: params }).map((res: Response) => {
+    return this.http.get(url, { params: params }).map((res: Response) => {
       
       return JSON.parse(xml2json(res.text(), ''));
       //return res.text();
@@ -225,13 +217,13 @@ export class AppService {
   
   getFromAleph(set_no: string, no_records:string){
     var url = this.state.config['context'] + 'aleph';
-    let params: URLSearchParams = new URLSearchParams();
+    let params: HttpParams = new HttpParams();
         
-    params.set('op', 'present');
-    params.set('format', 'marc');
-    params.set('set_no', set_no);
-    params.set('set_entry', '1-' +no_records);
-    return this.http.get(url, { search: params }).map((res: Response) => {
+    params = params.set('op', 'present');
+    params = params.set('format', 'marc');
+    params = params.set('set_no', set_no);
+    params = params.set('set_entry', '1-' +no_records);
+    return this.http.get(url, { params: params }).map((res: Response) => {
       
       return JSON.parse(xml2json(res.text(), ''));
       //return res.text();
@@ -245,97 +237,92 @@ export class AppService {
   }
   
   getDigKnihovnu(nazev: string): Observable<any> {
-    let params: URLSearchParams = new URLSearchParams();
-    params.set('q', 'nazev:"' + nazev +'"');
-    params.set('rows', '1');
+    let params: HttpParams = new HttpParams();
+    params = params.set('q', 'nazev:"' + nazev +'"');
+    params = params.set('rows', '1');
     var url = this.state.config['context'] + 'search/digknihovny/select';
-    return this.http.get(url, { search: params }).map(res => {
-      return res.json()['response']['docs'][0];
+    return this.http.get(url, { params: params }).map(res => {
+      return res['response']['docs'][0];
     });
   }
   
   getDigKnihovnyCount(nazev: string): Observable<number> {
-    let params: URLSearchParams = new URLSearchParams();
-    params.set('q', 'digknihovna:"' + nazev +'"');
-    params.set('rows', '0');
+    let params: HttpParams = new HttpParams();
+    params = params.set('q', 'digknihovna:"' + nazev +'"');
+    params = params.set('rows', '0');
     var url = this.state.config['context'] + 'search/rdcz/select';
-    return this.http.get(url, { search: params }).map(res => {
-      return res.json()['response']['numFound'];
+    return this.http.get(url, { params: params }).map(res => {
+      return res['response']['numFound'];
     });
   }
 
   getDigKnihovny(): Observable<any> {
-    let params: URLSearchParams = new URLSearchParams();
-    params.set('q', '*');
-    params.set('rows', '1000');
+    let params: HttpParams = new HttpParams();
+    params = params.set('q', '*');
+    params = params.set('rows', '1000');
 //    params.set('sort', 'nazev asc');
     var url = this.state.config['context'] + 'search/digknihovny/select';
-    return this.http.get(url, { search: params }).map(res => {
-      return res.json()['response']['docs'];
+    return this.http.get(url, { params: params }).map(res => {
+      return res['response']['docs'];
     });
   }
 
   getLists(): Observable<any> {
-    let params: URLSearchParams = new URLSearchParams();
-    params.set('q', '*');
+    let params: HttpParams = new HttpParams();
+    params = params.set('q', '*');
     let classes = 'dummyvalue'
     for (let i in this.state.config['facets']) {
       if(this.state.config['facets'][i]['classname']){
         classes += ' OR classname:"' + this.state.config['facets'][i]['classname'] + '"'
       }
     }
-    params.append('fq', classes);
-    params.set('rows', '1000');
+    params = params.append('fq', classes);
+    params = params.set('rows', '1000');
     var url = this.state.config['context'] + 'search/lists/select';
-    return this.http.get(url, { search: params }).map(res => {
-      return res.json()['response']['docs'];
+    return this.http.get(url, { params: params }).map(res => {
+      return res['response']['docs'];
     });
   }
   
   getEditablePages(): Observable<any>{
     var url = 'texts?action=LIST&lang=' + this.state.currentLang;
 
-    return this.http.get(url).map((response: Response) => {
-      return response.json();
-    }).catch(error => { return Observable.of('error gettting content: ' + error); });
+    return this.http.get(url);
   }
   
   
   getText(id: string): Observable<string> {
     var url = 'texts?action=LOAD&id=' + id + '&lang=' + this.state.currentLang;
-    return this.http.get(url).map((response: Response) => {
-      return response.text();
+    return this.http.get(url, {responseType: 'text'}).map((response) => {
+      return response;
     }).catch(error => { return Observable.of('error gettting content: ' + error); });
   }
 
   saveText(id: string, text: string): Observable<string> {
     var url = 'texts?action=SAVE&id=' + id + '&lang=' + this.state.currentLang;
 
-    let headers = new Headers({ 'Content-Type': 'text/plain;charset=UTF-8' });
-    let options = new RequestOptions({ headers: headers });
+    let headers = new HttpHeaders({ 'Content-Type': 'text/plain;charset=UTF-8' });
+    const options = { headers: headers };
     
     //we should convert back routerlink= to routerLink=
     
     let ctext = text.replace(/routerlink/g, 'routerLink');//.replace(/fragment/g, '[fragment]');
 
-    return this.http.post(url, ctext, options)
-      .map((response: Response) => {
-        return response.json();
-
-      }).catch(error => { return Observable.of('error saving content: ' + error); });
+    return this.http.post<string>(url, ctext, options);
+    // .catch(error => { return Observable.of('error saving content: ' + error); });
 
   }
 
   saveMenu(menu: any): Observable<string> {
-    let params: URLSearchParams = new URLSearchParams();
-    params.append('action', 'SAVEMENU');
-    params.append('menu', JSON.stringify(this.state.info_menu));
+    let params: HttpParams = new HttpParams()
+      .append('action', 'SAVEMENU')
+      .append('menu', JSON.stringify(this.state.info_menu));
     var url = 'texts';
-    return this.http.get(url, { search: params })
-      .map((response: Response) => {
-        return response.json();
-
-      }).catch(error => { return Observable.of('error saving content: ' + error); });
+    return this.http.get<string>(url, { params: params });
+//      .map((response: Response) => {
+//        return response;
+//
+//      }).catch(error => { return Observable.of('error saving content: ' + error); });
 
   }
 
@@ -370,12 +357,12 @@ export class AppService {
 
   doLogin() {
     var url = 'login'
-    var params = new URLSearchParams();
-    params.set('user', this.state.loginuser);
-    params.set('pwd', this.state.loginpwd);
-    params.set('action', 'LOGIN');
-    return this.http.get(url, { search: params }).map(res => {
-      return res.json();
+    var params = new HttpParams()
+    .set('user', this.state.loginuser)
+    .set('pwd', this.state.loginpwd)
+    .set('action', 'LOGIN');
+    return this.http.get(url, { params: params }).map(res => {
+      return res;
     }, error => {
       console.log('error : ' + error);
     });
@@ -397,11 +384,8 @@ export class AppService {
 
     var url = 'login';
     //console.log(this.loginuser, this.loginpwd, url);
-    var params = new URLSearchParams();
-    params.set('action', 'LOGOUT');
-    return this.http.get(url, { search: params }).map(res => {
-      return res.json();
-    });
+    var params = new HttpParams().set('action', 'LOGOUT');
+    return this.http.get(url, { params: params });
 
   }
 
